requirements.txt
pyyaml>=6.0
openai==0.27.8
matplotlib==3.5.3
pandas==1.5.3 